agent_api/pyproject.toml
[project]
name = "agent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "certifi==2024.8.30",
    "click==8.1.7",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.5",
    "fastapi==0.115.3",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.10",
    "jinja2==3.1.4",
    "jiter==0.6.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "openai==1.52.2",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pygments==2.18.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "pyyaml==6.0.2",
    "rich==13.9.3",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.41.0",
    "tqdm==4.66.5",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "uvicorn==0.32.0",
    "uvloop==0.21.0",
    "watchfiles==0.24.0",
    "websockets==13.1",
]
